SDL2: Fixed viewport on PSP by offset workaround
To resolve the problem of viewport just doesn't work as expected on PSP, I implemented the workaround with scissors and manual offsets to vertices that resolves the problem. Once genuine way to do that via PSP SDK will be figured, then this code can be adjusted. I used some pieces from the Vita's render code.